projects: cast float arguments to double in pr_info calls
Variadic float arguments are promoted to double anyway; make the
conversion explicit to silence -Wdouble-promotion in the affected
project examples.

+++ projects/eval-adis1654x/src/examples/basic_example/basic_example.c
@@ -131,7 +131,7 @@ int example_main()
 			goto exit;
 
 		for (uint8_t i = 0; i < 7; i++)
-			pr_info("%s %.5f %s \n", output_data[i], val[i] * output_scale[i],
+			pr_info("%s %.5f %s \n", output_data[i], (double)(val[i] * output_scale[i]),
 				output_unit[i]);
 	}
